+++ core/src/main/kotlin/net/corda/core/Utils.kt
@@ -1,49 +1,18 @@
-// TODO Move out the Kotlin specific stuff into a separate file
 @file:JvmName("Utils")
 
 package net.corda.core
 
-import com.google.common.base.Throwables
-import com.google.common.io.ByteStreams
 import com.google.common.util.concurrent.*
-import net.corda.core.crypto.SecureHash
-import net.corda.core.crypto.sha256
-import net.corda.core.internal.createDirectories
-import net.corda.core.internal.div
-import net.corda.core.internal.write
 import org.slf4j.Logger
 import rx.Observable
 import rx.Observer
-import rx.subjects.PublishSubject
-import rx.subjects.UnicastSubject
-import java.io.*
-import java.math.BigDecimal
-import java.nio.file.Files
-import java.nio.file.Path
 import java.time.Duration
 import java.util.concurrent.CompletableFuture
 import java.util.concurrent.ExecutionException
 import java.util.concurrent.Future
 import java.util.concurrent.TimeUnit
-import java.util.zip.Deflater
-import java.util.zip.ZipEntry
-import java.util.zip.ZipInputStream
-import java.util.zip.ZipOutputStream
 
-// TODO: Review by EOY2016 if we ever found these utilities helpful.
-val Int.bd: BigDecimal get() = BigDecimal(this)
-val Double.bd: BigDecimal get() = BigDecimal(this)
-val String.bd: BigDecimal get() = BigDecimal(this)
-val Long.bd: BigDecimal get() = BigDecimal(this)
-
-fun String.abbreviate(maxWidth: Int): String = if (length <= maxWidth) this else take(maxWidth - 1) + "…"
-
-/** Like the + operator but throws an exception in case of integer overflow. */
-infix fun Int.checkedAdd(b: Int) = Math.addExact(this, b)
-
-/** Like the + operator but throws an exception in case of integer overflow. */
-@Suppress("unused")
-infix fun Long.checkedAdd(b: Long) = Math.addExact(this, b)
+// TODO Delete this file once the Future stuff is out of here
 
 /** Same as [Future.get] but with a more descriptive name, and doesn't throw [ExecutionException], instead throwing its cause */
 fun <T> Future<T>.getOrThrow(timeout: Duration? = null): T {
@@ -92,163 +61,6 @@ fun <A> ListenableFuture<out A>.toObservable(): Observable<A> {
     }
 }
 
-/** Returns the index of the given item or throws [IllegalArgumentException] if not found. */
-fun <T> List<T>.indexOfOrThrow(item: T): Int {
-    val i = indexOf(item)
-    require(i != -1)
-    return i
-}
-
-/**
- * Returns the single element matching the given [predicate], or `null` if element was not found,
- * or throws if more than one element was found.
- */
-fun <T> Iterable<T>.noneOrSingle(predicate: (T) -> Boolean): T? {
-    var single: T? = null
-    for (element in this) {
-        if (predicate(element)) {
-            if (single == null) {
-                single = element
-            } else throw IllegalArgumentException("Collection contains more than one matching element.")
-        }
-    }
-    return single
-}
-
-/** Returns single element, or `null` if element was not found, or throws if more than one element was found. */
-fun <T> Iterable<T>.noneOrSingle(): T? {
-    var single: T? = null
-    for (element in this) {
-        if (single == null) {
-            single = element
-        } else throw IllegalArgumentException("Collection contains more than one matching element.")
-    }
-    return single
-}
-
-/** Returns a random element in the list, or null if empty */
-fun <T> List<T>.randomOrNull(): T? {
-    if (size <= 1) return firstOrNull()
-    val randomIndex = (Math.random() * size).toInt()
-    return get(randomIndex)
-}
-
-/** Returns a random element in the list matching the given predicate, or null if none found */
-fun <T> List<T>.randomOrNull(predicate: (T) -> Boolean) = filter(predicate).randomOrNull()
-
-inline fun elapsedTime(block: () -> Unit): Duration {
-    val start = System.nanoTime()
-    block()
-    val end = System.nanoTime()
-    return Duration.ofNanos(end - start)
-}
-
-// TODO: Add inline back when a new Kotlin version is released and check if the java.lang.VerifyError
-// returns in the IRSSimulationTest. If not, commit the inline back.
-fun <T> logElapsedTime(label: String, logger: Logger? = null, body: () -> T): T {
-    // Use nanoTime as it's monotonic.
-    val now = System.nanoTime()
-    try {
-        return body()
-    } finally {
-        val elapsed = Duration.ofNanos(System.nanoTime() - now).toMillis()
-        if (logger != null)
-            logger.info("$label took $elapsed msec")
-        else
-            println("$label took $elapsed msec")
-    }
-}
-
-fun <T> Logger.logElapsedTime(label: String, body: () -> T): T = logElapsedTime(label, this, body)
-
-/**
- * Given a path to a zip file, extracts it to the given directory.
- */
-fun extractZipFile(zipFile: Path, toDirectory: Path) = extractZipFile(Files.newInputStream(zipFile), toDirectory)
-
-/**
- * Given a zip file input stream, extracts it to the given directory.
- */
-fun extractZipFile(inputStream: InputStream, toDirectory: Path) {
-    val normalisedDirectory = toDirectory.normalize().createDirectories()
-    ZipInputStream(BufferedInputStream(inputStream)).use {
-        while (true) {
-            val e = it.nextEntry ?: break
-            val outPath = (normalisedDirectory / e.name).normalize()
-
-            // Security checks: we should reject a zip that contains tricksy paths that try to escape toDirectory.
-            check(outPath.startsWith(normalisedDirectory)) { "ZIP contained a path that resolved incorrectly: ${e.name}" }
-
-            if (e.isDirectory) {
-                outPath.createDirectories()
-                continue
-            }
-            outPath.write { out ->
-                ByteStreams.copy(it, out)
-            }
-            it.closeEntry()
-        }
-    }
-}
-
-/** Convert a [ByteArrayOutputStream] to [InputStreamAndHash]. */
-fun ByteArrayOutputStream.getInputStreamAndHash(baos: ByteArrayOutputStream): InputStreamAndHash {
-    val bytes = baos.toByteArray()
-    return InputStreamAndHash(ByteArrayInputStream(bytes), bytes.sha256())
-}
-
-data class InputStreamAndHash(val inputStream: InputStream, val sha256: SecureHash.SHA256) {
-    companion object {
-        /**
-         * Get a valid InputStream from an in-memory zip as required for some tests. The zip consists of a single file
-         * called "z" that contains the given content byte repeated the given number of times.
-         * Note that a slightly bigger than numOfExpectedBytes size is expected.
-         */
-        @Throws(IllegalArgumentException::class)
-        @JvmStatic
-        fun createInMemoryTestZip(numOfExpectedBytes: Int, content: Byte): InputStreamAndHash {
-            require(numOfExpectedBytes > 0)
-            val baos = ByteArrayOutputStream()
-            ZipOutputStream(baos).use { zos ->
-                val arraySize = 1024
-                val bytes = ByteArray(arraySize) { content }
-                val n = (numOfExpectedBytes - 1) / arraySize + 1 // same as Math.ceil(numOfExpectedBytes/arraySize).
-                zos.setLevel(Deflater.NO_COMPRESSION)
-                zos.putNextEntry(ZipEntry("z"))
-                for (i in 0 until n) {
-                    zos.write(bytes, 0, arraySize)
-                }
-                zos.closeEntry()
-            }
-            return baos.getInputStreamAndHash(baos)
-        }
-    }
-}
-
-// TODO: Generic csv printing utility for clases.
-
-val Throwable.rootCause: Throwable get() = Throwables.getRootCause(this)
-
-/**
- * Returns an Observable that buffers events until subscribed.
- * @see UnicastSubject
- */
-fun <T> Observable<T>.bufferUntilSubscribed(): Observable<T> {
-    val subject = UnicastSubject.create<T>()
-    val subscription = subscribe(subject)
-    return subject.doOnUnsubscribe { subscription.unsubscribe() }
-}
-
-/**
- * Copy an [Observer] to multiple other [Observer]s.
- */
-fun <T> Observer<T>.tee(vararg teeTo: Observer<T>): Observer<T> {
-    val subject = PublishSubject.create<T>()
-    subject.subscribe(this)
-    teeTo.forEach { subject.subscribe(it) }
-    return subject
-}
-
 /**
  * Returns a [ListenableFuture] bound to the *first* item emitted by this Observable. The future will complete with a
  * NoSuchElementException if no items are emitted or any other error thrown by the Observable. If it's cancelled then
@@ -273,15 +85,3 @@ private class ObservableToFuture<T>(observable: Observable<T>) : AbstractFuture<
 
     override fun onCompleted() {}
 }
-
-/** Return the sum of an Iterable of [BigDecimal]s. */
-fun Iterable<BigDecimal>.sum(): BigDecimal = fold(BigDecimal.ZERO) { a, b -> a + b }
-
-fun <T> Class<T>.checkNotUnorderedHashMap() {
-    if (HashMap::class.java.isAssignableFrom(this) && !LinkedHashMap::class.java.isAssignableFrom(this)) {
-        throw NotSerializableException("Map type $this is unstable under iteration. Suggested fix: use LinkedHashMap instead.")
-    }
-}
-
-fun Class<*>.requireExternal(msg: String = "Internal class")
-        = require(!name.startsWith("net.corda.node.") && !name.contains(".internal.")) { "$msg: $name" }
