+++ core/src/main/kotlin/net/corda/core/internal/InternalUtils.kt
@@ -1,16 +1,36 @@
 package net.corda.core.internal
 
+import com.google.common.base.Throwables
+import net.corda.core.crypto.SecureHash
+import net.corda.core.crypto.sha256
+import org.slf4j.Logger
+import rx.Observable
+import rx.Observer
+import rx.subjects.PublishSubject
+import rx.subjects.UnicastSubject
+import java.io.ByteArrayInputStream
+import java.io.ByteArrayOutputStream
 import java.io.InputStream
 import java.io.OutputStream
 import java.lang.reflect.Field
+import java.math.BigDecimal
 import java.nio.charset.Charset
 import java.nio.charset.StandardCharsets.UTF_8
 import java.nio.file.*
 import java.nio.file.attribute.FileAttribute
 import java.time.Duration
 import java.time.temporal.Temporal
+import java.util.*
+import java.util.Spliterator.*
+import java.util.stream.IntStream
 import java.util.stream.Stream
+import java.util.stream.StreamSupport
+import java.util.zip.Deflater
+import java.util.zip.ZipEntry
+import java.util.zip.ZipOutputStream
 import kotlin.reflect.KClass
+
+inline val Throwable.rootCause: Throwable get() = Throwables.getRootCause(this)
 
 infix fun Temporal.until(endExclusive: Temporal): Duration = Duration.between(this, endExclusive)
 
@@ -23,6 +43,50 @@ operator fun Duration.times(multiplicand: Long): Duration = multipliedBy(multipl
  */
 operator fun Path.div(other: String): Path = resolve(other)
 operator fun String.div(other: String): Path = Paths.get(this) / other
+
+/**
+ * Returns the single element matching the given [predicate], or `null` if the collection is empty, or throws exception
+ * if more than one element was found.
+ */
+inline fun <T> Iterable<T>.noneOrSingle(predicate: (T) -> Boolean): T? {
+    val iterator = iterator()
+    for (single in iterator) {
+        if (predicate(single)) {
+            while (iterator.hasNext()) {
+                if (predicate(iterator.next())) throw IllegalArgumentException("Collection contains more than one matching element.")
+            }
+            return single
+        }
+    }
+    return null
+}
+
+/**
+ * Returns the single element, or `null` if the list is empty, or throws an exception if it has more than one element.
+ */
+fun <T> List<T>.noneOrSingle(): T? {
+    return when (size) {
+        0 -> null
+        1 -> this[0]
+        else -> throw IllegalArgumentException("List has more than one element.")
+    }
+}
+
+/** Returns a random element in the list, or `null` if empty */
+fun <T> List<T>.randomOrNull(): T? {
+    return when (size) {
+        0 -> null
+        1 -> this[0]
+        else -> this[(Math.random() * size).toInt()]
+    }
+}
+
+/** Returns the index of the given item or throws [IllegalArgumentException] if not found. */
+fun <T> List<T>.indexOfOrThrow(item: T): Int {
+    val i = indexOf(item)
+    require(i != -1)
+    return i
+}
 
 fun Path.createDirectory(vararg attrs: FileAttribute<*>): Path = Files.createDirectory(this, *attrs)
 fun Path.createDirectories(vararg attrs: FileAttribute<*>): Path = Files.createDirectories(this, *attrs)
@@ -53,6 +117,112 @@ fun Path.readAllLines(charset: Charset = UTF_8): List<String> = Files.readAllLin
 fun Path.writeLines(lines: Iterable<CharSequence>, charset: Charset = UTF_8, vararg options: OpenOption): Path = Files.write(this, lines, charset, *options)
 
 fun InputStream.copyTo(target: Path, vararg options: CopyOption): Long = Files.copy(this, target, *options)
+
+fun String.abbreviate(maxWidth: Int): String = if (length <= maxWidth) this else take(maxWidth - 1) + "…"
+
+/** Return the sum of an Iterable of [BigDecimal]s. */
+fun Iterable<BigDecimal>.sum(): BigDecimal = fold(BigDecimal.ZERO) { a, b -> a + b }
+
+/**
+ * Returns an Observable that buffers events until subscribed.
+ * @see UnicastSubject
+ */
+fun <T> Observable<T>.bufferUntilSubscribed(): Observable<T> {
+    val subject = UnicastSubject.create<T>()
+    val subscription = subscribe(subject)
+    return subject.doOnUnsubscribe { subscription.unsubscribe() }
+}
+
+/** Copy an [Observer] to multiple other [Observer]s. */
+fun <T> Observer<T>.tee(vararg teeTo: Observer<T>): Observer<T> {
+    val subject = PublishSubject.create<T>()
+    subject.subscribe(this)
+    teeTo.forEach { subject.subscribe(it) }
+    return subject
+}
+
+/** Executes the given code block and returns a [Duration] of how long it took to execute in nanosecond precision. */
+inline fun elapsedTime(block: () -> Unit): Duration {
+    val start = System.nanoTime()
+    block()
+    val end = System.nanoTime()
+    return Duration.ofNanos(end - start)
+}
+
+
+fun <T> Logger.logElapsedTime(label: String, body: () -> T): T = logElapsedTime(label, this, body)
+
+// TODO: Add inline back when a new Kotlin version is released and check if the java.lang.VerifyError
+// returns in the IRSSimulationTest. If not, commit the inline back.
+fun <T> logElapsedTime(label: String, logger: Logger? = null, body: () -> T): T {
+    // Use nanoTime as it's monotonic.
+    val now = System.nanoTime()
+    try {
+        return body()
+    } finally {
+        val elapsed = Duration.ofNanos(System.nanoTime() - now).toMillis()
+        if (logger != null)
+            logger.info("$label took $elapsed msec")
+        else
+            println("$label took $elapsed msec")
+    }
+}
+
+/** Convert a [ByteArrayOutputStream] to [InputStreamAndHash]. */
+fun ByteArrayOutputStream.toInputStreamAndHash(): InputStreamAndHash {
+    val bytes = toByteArray()
+    return InputStreamAndHash(ByteArrayInputStream(bytes), bytes.sha256())
+}
+
+data class InputStreamAndHash(val inputStream: InputStream, val sha256: SecureHash.SHA256) {
+    companion object {
+        /**
+         * Get a valid InputStream from an in-memory zip as required for some tests. The zip consists of a single file
+         * called "z" that contains the given content byte repeated the given number of times.
+         * Note that a slightly bigger than numOfExpectedBytes size is expected.
+         */
+        fun createInMemoryTestZip(numOfExpectedBytes: Int, content: Byte): InputStreamAndHash {
+            require(numOfExpectedBytes > 0)
+            val baos = ByteArrayOutputStream()
+            ZipOutputStream(baos).use { zos ->
+                val arraySize = 1024
+                val bytes = ByteArray(arraySize) { content }
+                val n = (numOfExpectedBytes - 1) / arraySize + 1 // same as Math.ceil(numOfExpectedBytes/arraySize).
+                zos.setLevel(Deflater.NO_COMPRESSION)
+                zos.putNextEntry(ZipEntry("z"))
+                for (i in 0 until n) {
+                    zos.write(bytes, 0, arraySize)
+                }
+                zos.closeEntry()
+            }
+            return baos.toInputStreamAndHash()
+        }
+    }
+}
+
+fun IntIterator.toJavaIterator(): PrimitiveIterator.OfInt {
+    return object : PrimitiveIterator.OfInt {
+        override fun nextInt() = this@toJavaIterator.nextInt()
+        override fun hasNext() = this@toJavaIterator.hasNext()
+        override fun remove() = throw UnsupportedOperationException("remove")
+    }
+}
+
+private fun IntProgression.toSpliterator(): Spliterator.OfInt {
+    val spliterator = Spliterators.spliterator(
+            iterator().toJavaIterator(),
+            (1 + (last - first) / step).toLong(),
+            SUBSIZED or IMMUTABLE or NONNULL or SIZED or ORDERED or SORTED or DISTINCT
+    )
+    return if (step > 0) spliterator else object : Spliterator.OfInt by spliterator {
+        override fun getComparator() = Comparator.reverseOrder<Int>()
+    }
+}
+
+fun IntProgression.stream(parallel: Boolean = false): IntStream = StreamSupport.intStream(toSpliterator(), parallel)
+
+@Suppress("UNCHECKED_CAST") // When toArray has filled in the array, the component type is no longer T? but T (that may itself be nullable).
+inline fun <reified T> Stream<out T>.toTypedArray() = toArray { size -> arrayOfNulls<T>(size) } as Array<T>
 
 fun <T> Class<T>.castIfPossible(obj: Any): T? = if (isInstance(obj)) cast(obj) else null
