+++ node-api/src/main/kotlin/net/corda/nodeapi/RPCStructures.kt
@@ -5,11 +5,10 @@ package net.corda.nodeapi
 import com.esotericsoftware.kryo.Registration
 import com.esotericsoftware.kryo.Serializer
 import com.google.common.util.concurrent.ListenableFuture
-import net.corda.core.requireExternal
+import net.corda.core.CordaRuntimeException
 import net.corda.core.serialization.*
 import net.corda.core.toFuture
 import net.corda.core.toObservable
-import net.corda.core.CordaRuntimeException
 import net.corda.nodeapi.config.OldConfig
 import rx.Observable
 import java.io.InputStream
@@ -72,5 +71,9 @@ class RPCKryo(observableSerializer: Serializer<Observable<Any>>, whitelist: Clas
         }
         type.requireExternal("RPC not allowed to deserialise internal classes")
         return super.getRegistration(type)
+    }
+
+    private fun Class<*>.requireExternal(msg: String) {
+        require(!name.startsWith("net.corda.node.") && !name.contains(".internal.")) { "$msg: $name" }
     }
 }
