+++ node-api/src/main/kotlin/net/corda/nodeapi/config/ConfigUtilities.kt
@@ -2,7 +2,7 @@ package net.corda.nodeapi.config
 
 import com.typesafe.config.Config
 import com.typesafe.config.ConfigUtil
-import net.corda.core.noneOrSingle
+import net.corda.core.internal.noneOrSingle
 import net.corda.core.utilities.validateX500Name
 import net.corda.core.utilities.NetworkHostAndPort
 import net.corda.core.utilities.parseNetworkHostAndPort
