+++ node/src/main/kotlin/net/corda/node/internal/NodeStartup.kt
@@ -6,11 +6,9 @@ import joptsimple.OptionException
 import net.corda.core.*
 import net.corda.core.crypto.commonName
 import net.corda.core.crypto.orgName
-import net.corda.core.internal.createDirectories
-import net.corda.core.internal.div
+import net.corda.core.internal.*
 import net.corda.node.VersionInfo
 import net.corda.core.node.services.ServiceInfo
-import net.corda.core.internal.Emoji
 import net.corda.core.utilities.loggerFor
 import net.corda.node.ArgsParser
 import net.corda.node.CmdLineOptions
