+++ node/src/main/kotlin/net/corda/node/services/persistence/DBTransactionStorage.kt
@@ -1,7 +1,7 @@
 package net.corda.node.services.persistence
 
 import com.google.common.annotations.VisibleForTesting
-import net.corda.core.bufferUntilSubscribed
+import net.corda.core.internal.bufferUntilSubscribed
 import net.corda.core.crypto.SecureHash
 import net.corda.core.messaging.DataFeed
 import net.corda.core.serialization.SingletonSerializeAsToken
