+++ samples/simm-valuation-demo/src/main/kotlin/net/corda/vega/portfolio/Portfolio.kt
@@ -3,11 +3,12 @@ package net.corda.vega.portfolio
 import net.corda.client.rpc.notUsed
 import net.corda.core.contracts.*
 import net.corda.core.identity.Party
+import net.corda.core.internal.sum
 import net.corda.core.messaging.CordaRPCOps
 import net.corda.core.node.ServiceHub
-import net.corda.core.sum
 import net.corda.vega.contracts.IRSState
 import net.corda.vega.contracts.SwapData
+import java.math.BigDecimal
 import java.time.LocalDate
 
 /**
@@ -22,7 +23,7 @@ data class Portfolio(private val tradeStateAndRefs: List<StateAndRef<IRSState>>,
     val swaps: List<SwapData> by lazy { trades.map { it.swap } }
     val refs: List<StateRef> by lazy { tradeStateAndRefs.map { it.ref } }
 
-    fun getNotionalForParty(party: Party) = trades.map { it.swap.getLegForParty(party).notional }.sum()
+    fun getNotionalForParty(party: Party): BigDecimal = trades.map { it.swap.getLegForParty(party).notional }.sum()
 
     fun update(curTrades: List<StateAndRef<IRSState>>): Portfolio {
         return copy(tradeStateAndRefs = curTrades)
