+++ test-utils/src/main/kotlin/net/corda/testing/node/TestClock.kt
@@ -4,6 +4,7 @@ import net.corda.core.internal.until
 import net.corda.core.serialization.SerializeAsToken
 import net.corda.core.serialization.SerializeAsTokenContext
 import net.corda.core.serialization.SingletonSerializationToken.Companion.singletonSerializationToken
+import net.corda.core.internal.until
 import net.corda.node.utilities.MutableClock
 import java.time.Clock
 import java.time.Duration
